# apps/backend/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/projects/mine": {
      "get": {
        "summary": "List My Projects",
        "tags": [
          "Project"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "operationId": "listMine",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "supplierId": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "imageUrl": {
                        "type": "string",
                        "format": "uri"
                      },
                      "locationLat": {
                        "type": "number"
                      },
                      "locationLng": {
                        "type": "number"
                      },
                      "area": {
                        "type": "number"
                      },
                      "contractAddress": {
                        "type": "string",
                        "nullable": true
                      },
                      "tokenSymbol": {
                        "type": "string",
                        "nullable": true
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "DRAFT",
                          "PENDING",
                          "APPROVED",
                          "DEPLOYED",
                          "REJECTED"
                        ]
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "supplierId",
                      "title",
                      "description",
                      "imageUrl",
                      "locationLat",
                      "locationLng",
                      "area",
                      "status",
                      "createdAt"
                    ]
                  }
                }
              }
            },
            "headers": {
              "X-Total-Count": {
                "description": "Total number of items matching the query",
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              },
              "X-Page": {
                "description": "Current page number (1-based)",
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "X-Limit": {
                "description": "Page size (items per page)",
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/projects/{id}/approve": {
      "post": {
        "summary": "Approve Project (Admin)",
        "tags": [
          "Project"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "approve",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "supplierId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "imageUrl": {
                      "type": "string",
                      "format": "uri"
                    },
                    "locationLat": {
                      "type": "number"
                    },
                    "locationLng": {
                      "type": "number"
                    },
                    "area": {
                      "type": "number"
                    },
                    "contractAddress": {
                      "type": "string",
                      "nullable": true
                    },
                    "tokenSymbol": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "PENDING",
                        "APPROVED",
                        "DEPLOYED",
                        "REJECTED"
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "supplierId",
                    "title",
                    "description",
                    "imageUrl",
                    "locationLat",
                    "locationLng",
                    "area",
                    "status",
                    "createdAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "Ledgera REST API - Documentation",
    "description": "Ledgera REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}